New Chart, Fullscreen, Optimization
First, I'm really excited to introduce the hitherto most complicated
chart, the bar chart. It calculates some good ranges and looks how much
Solves in these Ranges are.

Then I also added the Fullscreen Toggle. With a doubleclick, you can now
toggle the Fullscreen mode of the Graphs.

And at last but not least, I optimizated the code. The part that will be
on every graph the same is now stored in the global default highchart
options and to add a new graph you now just have to set the some options
which are different for most of the graphs.

Just simplyfiying life...